Set elasticsearch client version to >=7.17,<8.0.0
The latest versions of the elasticsearch client have a compatibility
mode which can also be used for newer server versions; however, the other
way around does not work. Fixing the version to >=7.17,<8.0.0 is
therefore a save option for the near future.
This commit also includes a fix for the case where "resource_status" is
not part of `resource_attributes`, which occasionally caused crashes. If
this key is not present, `resource_status` will be set to an empty
string.