Add error handling when updating non-existing record
It can happen that a drone does not reach the AvailableState but only
the DownState. In this case, updating the record in Auditor does not
work, because no record has been created. In this case, Auditor returns
a 400 BAD REQUEST. This commit adds error handling for this specific
case. Other exceptions will not be handled.